Consolidate variant manager module
